Create shared folder (WIP)
Establish a folder named "shared" that contains files that are to be
shared between Magerun 1 and Magerun 2.

The directory structure inside "shared" mirrors the layout in the project
root to allow sharing of not only PHP code (src) but also build scripts
and any other kind of file.

Sharing of PHP files work via composer auto-loader. PSR-4 is preferred
over PSR-0 so that a file placed into the shared/src folder will be
preferred over the one in the src folder.

The phar-build works with a new file-set for shared.

Phpunit tests are similar, the test-suite is extended with the new
shared/tests directory.